JettonManager updates
