+++ packages/walletkit/src/core/JettonsManager.ts
@@ -14,10 +14,10 @@ import { LRUCache } from 'lru-cache';
 import type { EmulationTokenInfoMasters } from '../types/toncenter/emulation';
 import { globalLogger } from './Logger';
 import type { WalletKitEventEmitter } from '../types/emitter';
-import type { JettonInfo, JettonsAPI } from '../types/jettons';
+import type { JettonsAPI } from '../types/jettons';
 import { JettonError, JettonErrorCode } from '../types/jettons';
 import type { NetworkManager } from './NetworkManager';
-import type { Jetton } from '../api/models';
+import type { Jetton, JettonInfo } from '../api/models';
 import type { Network } from '../api/models';
 import { asMaybeAddressFriendly } from '../utils';
 
@@ -133,7 +133,7 @@ export class JettonsManager implements JettonsAPI {
                 }
 
                 const result: JettonInfo = {
-                    address: jetton.jetton,
+                    address: jetton.address,
                     name: tokenInfo?.name ?? '',
                     symbol: tokenInfo?.symbol ?? '',
                     description: tokenInfo?.description ?? '',
