+++ packages/walletkit/src/types/index.ts
@@ -32,8 +32,6 @@ export { DEFAULT_DURABLE_EVENTS_CONFIG } from './durableEvents';
 
 // Jettons types
 export type {
-    JettonInfo,
-    JettonVerification,
     AddressJetton,
     JettonBalance,
     JettonTransferParams,
